% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_backtranslate.R
\name{.aa_one_letter}
\alias{.aa_one_letter}
\title{Convert three-letter AA code to single-letter}
\usage{
.aa_one_letter(aa3)
}
\description{
Convert three-letter AA code to single-letter
}
\keyword{internal}
