% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_transcribe.R
\name{.build_c_notation}
\alias{.build_c_notation}
\title{Build a c. notation string from coordinates}
\usage{
.build_c_notation(tx_id, cds_pos, offset, is_utr5, is_utr3, ref, alt)
}
\description{
Build a c. notation string from coordinates
}
\keyword{internal}
