% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_transcribe.R
\name{.compute_cds_from_intronic}
\alias{.compute_cds_from_intronic}
\title{Compute CDS-relative position and offset for intronic genomic pos}
\usage{
.compute_cds_from_intronic(gpos, chr, exons_by_tx, tx_id, strand)
}
\description{
Compute CDS-relative position and offset for intronic genomic pos
}
\keyword{internal}
