% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_variants.R
\name{.compute_cds_pos}
\alias{.compute_cds_pos}
\title{Compute CDS-relative position}
\usage{
.compute_cds_pos(gpos, cds_start, cds_end, strand, region_type)
}
\arguments{
\item{gpos}{Genomic position (1-based).}

\item{cds_start}{Genomic coordinate of CDS start (ATG).}

\item{cds_end}{Genomic coordinate of CDS end (stop codon last base).}

\item{strand}{Strand ("+" or "-").}

\item{region_type}{Region type ("cds", "five_utr", "three_utr").}
}
\value{
Integer: positive for CDS, negative for 5'UTR,
NA for 3'UTR (stored separately, * notation added by formatter).
}
\description{
Maps a genomic coordinate to a CDS-relative position following
HGVS numbering conventions.
}
\keyword{internal}
