% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_parse.R
\name{.detect_hgvs_type}
\alias{.detect_hgvs_type}
\title{Detect HGVS variant type from the variant body string}
\usage{
.detect_hgvs_type(body)
}
\description{
Detect HGVS variant type from the variant body string
}
\keyword{internal}
