% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_backtranslate.R
\name{.enumerate_coding_variants}
\alias{.enumerate_coding_variants}
\title{Enumerate all possible coding variants for a protein change}
\usage{
.enumerate_coding_variants(protein_var, txdb, bsgenome, genetic_code)
}
\description{
Enumerate all possible coding variants for a protein change
}
\keyword{internal}
