% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_backtranslate.R
\name{.enumerate_missense}
\alias{.enumerate_missense}
\title{Enumerate missense codon changes}
\usage{
.enumerate_missense(pv, ref_cds, ref_protein, genetic_code, tx_id)
}
\description{
Enumerate missense codon changes
}
\keyword{internal}
