% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_backtranslate.R
\name{.enumerate_nonsense}
\alias{.enumerate_nonsense}
\title{Enumerate nonsense codon changes}
\usage{
.enumerate_nonsense(pv, ref_cds, ref_protein, genetic_code, tx_id)
}
\description{
Enumerate nonsense codon changes
}
\keyword{internal}
