% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_transcribe.R
\name{.extract_chromosome}
\alias{.extract_chromosome}
\title{Extract chromosome identifier from an accession}
\usage{
.extract_chromosome(accession)
}
\description{
Extract chromosome identifier from an accession
}
\keyword{internal}
