% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_extract.R
\name{.extract_variants_from_alignment}
\alias{.extract_variants_from_alignment}
\title{Extract HGVS variants from a pairwise alignment}
\usage{
.extract_variants_from_alignment(aln, accession, notation, start_pos)
}
\description{
Extract HGVS variants from a pairwise alignment
}
\keyword{internal}
