% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_normalize.R
\name{.format_single_pos}
\alias{.format_single_pos}
\title{Format a single HGVS coordinate}
\usage{
.format_single_pos(pos, offset, is_five_utr, is_three_utr)
}
\description{
Format a single HGVS coordinate
}
\keyword{internal}
