% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_variants.R
\name{.generate_region_variants}
\alias{.generate_region_variants}
\title{Generate variants for a single region}
\usage{
.generate_region_variants(
  gr,
  bsgenome,
  tx_id,
  cds_start,
  cds_end,
  strand,
  region_type
)
}
\arguments{
\item{gr}{A GRanges of region positions.}

\item{bsgenome}{BSgenome object.}

\item{tx_id}{Transcript ID.}

\item{cds_start}{CDS start genomic position.}

\item{cds_end}{CDS end genomic position.}

\item{strand}{Strand.}

\item{region_type}{One of "cds", "five_utr", "three_utr".}
}
\value{
data.frame of variants.
}
\description{
Generate variants for a single region
}
\keyword{internal}
