% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_variants.R
\name{.generate_splice_variants}
\alias{.generate_splice_variants}
\title{Generate splice site variants}
\usage{
.generate_splice_variants(splice_gr, bsgenome, tx_id, cds_start, strand)
}
\arguments{
\item{splice_gr}{GRanges of splice site positions from
transcript_structure.}

\item{bsgenome}{BSgenome object.}

\item{tx_id}{Transcript ID to filter.}

\item{cds_start}{CDS start genomic position.}

\item{strand}{Strand.}
}
\value{
data.frame of splice site variants.
}
\description{
Generate splice site variants
}
\keyword{internal}
