% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_translate.R
\name{.get_cds_sequence}
\alias{.get_cds_sequence}
\title{Retrieve coding sequence from BSgenome for a CDS GRanges}
\usage{
.get_cds_sequence(cds_gr, bsgenome, strand)
}
\description{
Retrieve coding sequence from BSgenome for a CDS GRanges
}
\keyword{internal}
