% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_convert.R
\name{.get_chr_accession}
\alias{.get_chr_accession}
\title{Get standard chromosome accession from assembly}
\usage{
.get_chr_accession(chr, assembly)
}
\description{
Get standard chromosome accession from assembly
}
\keyword{internal}
