% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_convert.R
\name{.hgvs_to_vcf_single}
\alias{.hgvs_to_vcf_single}
\title{Convert a single parsed HGVS to VCF}
\usage{
.hgvs_to_vcf_single(parsed, txdb, bsgenome)
}
\description{
Convert a single parsed HGVS to VCF
}
\keyword{internal}
