% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_liftover.R
\name{.liftover_single}
\alias{.liftover_single}
\title{Lift over a single g. variant}
\usage{
.liftover_single(parsed, chain, target_bsgenome)
}
\description{
Lift over a single g. variant
}
\keyword{internal}
