% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_extract.R
\name{.needleman_wunsch}
\alias{.needleman_wunsch}
\title{Needleman-Wunsch global alignment}
\usage{
.needleman_wunsch(
  seq1,
  seq2,
  match = 2,
  mismatch = -1,
  gap_open = -2,
  gap_extend = -1
)
}
\description{
Needleman-Wunsch global alignment
}
\keyword{internal}
