% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_normalize.R
\name{.normalize_variant}
\alias{.normalize_variant}
\title{Normalize a single parsed variant}
\usage{
.normalize_variant(parsed, bsgenome, txdb)
}
\description{
Normalize a single parsed variant
}
\keyword{internal}
