% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_extract.R
\name{.pairwise_align}
\alias{.pairwise_align}
\title{Simple pairwise sequence alignment (Needleman-Wunsch or
Smith-Waterman)}
\usage{
.pairwise_align(ref, obs, algorithm)
}
\description{
Simple pairwise sequence alignment (Needleman-Wunsch or
Smith-Waterman)
}
\keyword{internal}
