% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_backtranslate.R
\name{.parse_aa_three_letter_string}
\alias{.parse_aa_three_letter_string}
\title{Parse a string of concatenated three-letter amino acid codes}
\usage{
.parse_aa_three_letter_string(s)
}
\description{
Parse a string of concatenated three-letter amino acid codes
}
\keyword{internal}
