% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_parse.R
\name{.parse_deletion}
\alias{.parse_deletion}
\title{Parse deletion: e.g., 215delC, 215_217delTCA, 215+1delG}
\usage{
.parse_deletion(body)
}
\description{
Parse deletion: e.g., 215delC, 215_217delTCA, 215+1delG
}
\keyword{internal}
