% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_parse.R
\name{.parse_duplication}
\alias{.parse_duplication}
\title{Parse duplication: e.g., 215dupC, 215_217dup}
\usage{
.parse_duplication(body)
}
\description{
Parse duplication: e.g., 215dupC, 215_217dup
}
\keyword{internal}
