% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_parse.R
\name{.parse_frameshift}
\alias{.parse_frameshift}
\title{Parse frameshift: e.g., 215fs, 215delGfs*5}
\usage{
.parse_frameshift(body)
}
\description{
Parse frameshift: e.g., 215fs, 215delGfs*5
}
\keyword{internal}
