% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_parse.R
\name{.parse_insertion}
\alias{.parse_insertion}
\title{Parse insertion: e.g., 215_216insTAG, 215+1insG}
\usage{
.parse_insertion(body)
}
\description{
Parse insertion: e.g., 215_216insTAG, 215+1insG
}
\keyword{internal}
