% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_backtranslate.R
\name{.parse_protein_variant}
\alias{.parse_protein_variant}
\title{Parse a p. HGVS string (simplified parser for protein notation)}
\usage{
.parse_protein_variant(p_str)
}
\description{
Parse a p. HGVS string (simplified parser for protein notation)
}
\keyword{internal}
