% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_parse.R
\name{.parse_repeat}
\alias{.parse_repeat}
\title{Parse repeat variant: e.g., \code{215[3]}, 215_217(2)}
\usage{
.parse_repeat(body)
}
\description{
Parse repeat variant: e.g., \code{215[3]}, 215_217(2)
}
\keyword{internal}
