% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_parse.R
\name{.parse_single_pos}
\alias{.parse_single_pos}
\title{Parse a single HGVS coordinate (not a range)}
\usage{
.parse_single_pos(coord)
}
\description{
Parse a single HGVS coordinate (not a range)
}
\keyword{internal}
