% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_parse.R
\name{.parse_substitution}
\alias{.parse_substitution}
\title{Parse substitution: e.g., 215C>G, -14G>A, *32T>C, 453+1G>T}
\usage{
.parse_substitution(body)
}
\description{
Parse substitution: e.g., 215C>G, -14G>A, *32T>C, 453+1G>T
}
\keyword{internal}
