% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_parse.R
\name{.parse_variant_body}
\alias{.parse_variant_body}
\title{Parse position, reference, and alternate from variant body}
\usage{
.parse_variant_body(body, type)
}
\description{
Parse position, reference, and alternate from variant body
}
\keyword{internal}
