% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_variants.R
\name{.range_to_positions}
\alias{.range_to_positions}
\title{Expand GRanges to a vector of individual positions}
\usage{
.range_to_positions(gr, as_genomic = TRUE)
}
\arguments{
\item{gr}{A GRanges object.}

\item{as_genomic}{If TRUE, return genomic coordinates.
If FALSE, return sequential indices.}
}
\value{
Integer vector of positions.
}
\description{
Expand GRanges to a vector of individual positions
}
\keyword{internal}
