% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_normalize.R
\name{.rebuild_position}
\alias{.rebuild_position}
\title{Rebuild position string from position list}
\usage{
.rebuild_position(pos)
}
\description{
Rebuild position string from position list
}
\keyword{internal}
