% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_transcribe.R
\name{.resolve_intronic_position}
\alias{.resolve_intronic_position}
\title{Resolve intronic CDS position to genomic coordinates}
\usage{
.resolve_intronic_position(cds_pos, offset, cds_gr, exon_gr, strand)
}
\description{
Given a CDS exon boundary position and intronic offset,
compute the genomic position.
}
\keyword{internal}
