% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_transcribe.R
\name{.resolve_transcript}
\alias{.resolve_transcript}
\title{Resolve a transcript accession to a transcript ID in the TxDb}
\usage{
.resolve_transcript(accession, cds_by_tx, transcript_id)
}
\description{
Resolve a transcript accession to a transcript ID in the TxDb
}
\keyword{internal}
