% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_extract.R
\name{.smith_waterman}
\alias{.smith_waterman}
\title{Smith-Waterman local alignment}
\usage{
.smith_waterman(
  seq1,
  seq2,
  match = 2,
  mismatch = -1,
  gap_open = -2,
  gap_extend = -1
)
}
\description{
Smith-Waterman local alignment
}
\keyword{internal}
