% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_translate.R
\name{.standard_genetic_code}
\alias{.standard_genetic_code}
\title{Standard genetic code (DNA codons → single-letter amino acid)}
\usage{
.standard_genetic_code()
}
\description{
Standard genetic code (DNA codons → single-letter amino acid)
}
\keyword{internal}
