% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_validate.R
\name{.syntactic_checks}
\alias{.syntactic_checks}
\title{Syntactic validation rules}
\usage{
.syntactic_checks(parsed)
}
\arguments{
\item{parsed}{Parsed HGVS variant object.}
}
\value{
List with \code{valid} and \code{error} fields.
}
\description{
Checks HGVS grammar per HGVS nomenclature guidelines.
}
\keyword{internal}
