% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_normalize.R
\name{.three_prime_shift}
\alias{.three_prime_shift}
\title{Apply 3' shifting rule for indels in repetitive regions.}
\usage{
.three_prime_shift(parsed, bsgenome, txdb)
}
\description{
For deletions: shift position downstream as far as possible
while the deleted sequence still matches the reference.
}
\details{
For insertions: shift position downstream as far as possible
while the inserted base matches the following reference base.
}
\keyword{internal}
