% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_transcribe.R
\name{.transcribe_g_to_c}
\alias{.transcribe_g_to_c}
\title{Map g. variant to c. notation}
\usage{
.transcribe_g_to_c(parsed, txdb, bsgenome, transcript_id)
}
\description{
Uses TxDb to resolve a genomic position to CDS-relative coordinates.
}
\keyword{internal}
