% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_translate.R
\name{.translate_deletion}
\alias{.translate_deletion}
\title{Translate a deletion variant}
\usage{
.translate_deletion(parsed, ref_cds, ref_protein, cds_pos, genetic_code)
}
\description{
Translate a deletion variant
}
\keyword{internal}
