% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_translate.R
\name{.translate_delins}
\alias{.translate_delins}
\title{Translate a delins variant}
\usage{
.translate_delins(parsed, ref_cds, ref_protein, cds_pos, genetic_code)
}
\description{
Translate a delins variant
}
\keyword{internal}
