% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_translate.R
\name{.translate_duplication}
\alias{.translate_duplication}
\title{Translate a duplication variant}
\usage{
.translate_duplication(parsed, ref_cds, ref_protein, cds_pos, genetic_code)
}
\description{
Translate a duplication variant
}
\keyword{internal}
