% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_translate.R
\name{.translate_insertion}
\alias{.translate_insertion}
\title{Translate an insertion variant}
\usage{
.translate_insertion(parsed, ref_cds, ref_protein, cds_pos, genetic_code)
}
\description{
Translate an insertion variant
}
\keyword{internal}
