% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_translate.R
\name{.translate_sequence}
\alias{.translate_sequence}
\title{Translate a nucleotide sequence to protein using genetic code}
\usage{
.translate_sequence(nt_seq, genetic_code)
}
\description{
Translate a nucleotide sequence to protein using genetic code
}
\keyword{internal}
