% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_translate.R
\name{.translate_substitution}
\alias{.translate_substitution}
\title{Translate a substitution variant}
\usage{
.translate_substitution(parsed, ref_cds, ref_protein, cds_pos, genetic_code)
}
\description{
Translate a substitution variant
}
\keyword{internal}
