% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_translate.R
\name{.translate_variant}
\alias{.translate_variant}
\title{Translate a single parsed c. variant to p. notation}
\usage{
.translate_variant(parsed, txdb, bsgenome, genetic_code)
}
\description{
Translate a single parsed c. variant to p. notation
}
\keyword{internal}
