% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_normalize.R
\name{.trim_common_affixes}
\alias{.trim_common_affixes}
\title{Trim common prefix and suffix from reference and alternate
sequences for delins variants.}
\usage{
.trim_common_affixes(parsed)
}
\description{
Trim common prefix and suffix from reference and alternate
sequences for delins variants.
}
\keyword{internal}
