% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_normalize.R
\name{.try_rewrite_as_dup}
\alias{.try_rewrite_as_dup}
\title{Rewrite insertion as duplication when the inserted sequence matches
the adjacent reference sequence.}
\usage{
.try_rewrite_as_dup(parsed, bsgenome, txdb)
}
\description{
Rewrite insertion as duplication when the inserted sequence matches
the adjacent reference sequence.
}
\keyword{internal}
