% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_variants.R
\name{.validate_seqlevels}
\alias{.validate_seqlevels}
\title{Validate that chromosome names in the structure exist in the
BSgenome.}
\usage{
.validate_seqlevels(tx_struct, bsgenome)
}
\arguments{
\item{tx_struct}{Transcript structure object.}

\item{bsgenome}{BSgenome object.}
}
\value{
Logical: TRUE if all seqlevels are valid.
}
\description{
Validate that chromosome names in the structure exist in the
BSgenome.
}
\keyword{internal}
