% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_validate.R
\name{.verify_reference_allele}
\alias{.verify_reference_allele}
\title{Verify reference allele matches genome}
\usage{
.verify_reference_allele(parsed, txdb, bsgenome)
}
\description{
Verify reference allele matches genome
}
\keyword{internal}
