% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_convert.R
\name{hgvs_to_spdi}
\alias{hgvs_to_spdi}
\title{Convert HGVS to SPDI format}
\usage{
hgvs_to_spdi(hgvs_strings, assembly = "GRCh38")
}
\arguments{
\item{hgvs_strings}{Character vector of HGVS g. notation strings.}

\item{assembly}{Genome assembly name.}
}
\value{
Character vector of SPDI strings.
}
\description{
SPDI uses 0-based interbase coordinates:
seqid:position:deletion:insertion
}
\examples{
hgvs_to_spdi("NC_000001.11:g.123456A>G")
}
