% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_liftover.R
\name{liftover_c_hgvs}
\alias{liftover_c_hgvs}
\title{Lift over a c. variant by converting to genomic, lifting,
then converting back}
\usage{
liftover_c_hgvs(
  hgvs_c,
  source_txdb,
  source_bsgenome,
  chain_file,
  target_txdb,
  target_bsgenome
)
}
\arguments{
\item{hgvs_c}{A HGVS c. notation string.}

\item{source_txdb}{TxDb for the source assembly.}

\item{source_bsgenome}{BSgenome for the source assembly.}

\item{chain_file}{Path to chain file.}

\item{target_txdb}{TxDb for the target assembly.}

\item{target_bsgenome}{BSgenome for the target assembly.}
}
\value{
HGVS c. notation string in the target assembly, or NA.
}
\description{
Lift over a c. variant by converting to genomic, lifting,
then converting back
}
\examples{
if (requireNamespace("rtracklayer", quietly = TRUE)) {
    cat("rtracklayer available for c. liftover")
}

}
