% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hgvs_convert.R
\name{spdi_to_hgvs}
\alias{spdi_to_hgvs}
\title{Convert SPDI to HGVS g. notation}
\usage{
spdi_to_hgvs(spdi_strings)
}
\arguments{
\item{spdi_strings}{Character vector of SPDI strings.}
}
\value{
Character vector of HGVS g. notation strings.
}
\description{
Convert SPDI to HGVS g. notation
}
\examples{
spdi_to_hgvs("NC_000001.11:123455:A:G")

}
